Add Users to package and @Repository
