+++ src/main/java/com/example/kwbruunauktion/auktionSystem/entity/users/UserEconomy.java
@@ -1,21 +1,26 @@
-package com.example.kwbruunauktion.auktionSystem.entity;
+package com.example.kwbruunauktion.auktionSystem.entity.users;
 
+import com.example.kwbruunauktion.auktionSystem.entity.Ownership;
 import com.example.kwbruunauktion.security.entity.UserWithRoles;
 import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.OneToOne;
+import java.time.LocalDateTime;
 
 @Getter
 @Setter
 @ToString
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 
 @Entity
-public class UserAdmin extends UserWithRoles {
+public class UserEconomy extends UserWithRoles {
+
     @Column(length = 100, nullable = false)
     private String firstName;
 
@@ -27,4 +32,12 @@ public class UserAdmin extends UserWithRoles {
 
     @OneToOne(cascade = CascadeType.ALL)
     private Ownership ownership;
+
+    @CreationTimestamp
+    private LocalDateTime created;
+
+    @UpdateTimestamp
+    private LocalDateTime updated;
+
+
 }
