+++ src/main/java/com/example/kwbruunauktion/auktionSystem/repository/UserEconomyRepository.java
@@ -1,8 +1,9 @@
 package com.example.kwbruunauktion.auktionSystem.repository;
 
-import com.example.kwbruunauktion.auktionSystem.entity.SpecificCar;
-import com.example.kwbruunauktion.auktionSystem.entity.UserEconomy;
+import com.example.kwbruunauktion.auktionSystem.entity.users.UserEconomy;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface UserEconomyRepository extends JpaRepository<UserEconomy, Long> {
 }
