+++ src/main/java/com/example/kwbruunauktion/auktionSystem/repository/UserLeaserRepository.java
@@ -2,6 +2,8 @@ package com.example.kwbruunauktion.auktionSystem.repository;
 
 import com.example.kwbruunauktion.auktionSystem.entity.users.UserLeaser;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface UserLeaserRepository extends JpaRepository<UserLeaser, Long> {
 }
